requirements.txt
agate==1.6.3
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.5.1
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.0.0
colorama==0.4.5
commonmark==0.9.1
cryptography==36.0.2
dash==2.5.1
dash-bootstrap-components==1.2.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbt-core==1.2.0
dbt-extractor==0.4.1
dbt-snowflake==1.2.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.1.3
Flask-Compress==1.12
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
hologram==0.0.15
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keyring==23.7.0
leather==0.3.4
Logbook==1.5.3
MarkupSafe==2.0.1
mashumaro==2.9
matplotlib-inline==0.1.3
minimal-snowplow-tracker==0.0.2
mistune==0.8.4
msgpack==1.0.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
notebook==6.4.12
numpy==1.23.1
oscrypto==1.3.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parsedatetime==2.4
parso==0.8.3
pathlib==1.0.1
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.9.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pyarrow==6.0.0
pycparser==2.21
pycryptodomex==3.15.0
pydeck==0.7.1
Pygments==2.12.0
PyJWT==2.4.0
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytimeparse==1.1.8
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.6
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
rich==12.5.1
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
snowflake-connector-python==2.7.9
snowflake-snowpark-python==0.7.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
streamlit==1.10.0
tenacity==8.0.1
terminado==0.15.0
text-unidecode==1.3
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
validators==0.20.0
vega-datasets==0.9.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.1
zipp==3.8.1